// stylishlab-bridebox-frontend/src/api/generated/model/employeeResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Stylish Lab Bridebox - Salon Management API
 * API for managing salon services, employees, customers, sales, expenses, and reports
 * OpenAPI spec version: 1.0.0
 */

export interface EmployeeResponse {
  id?: number;
  fullName?: string;
  mobile?: string;
  joinDate?: string;
  profilePicture?: string;
  status?: string;
  username?: string;
  currentEmployeePercent?: number;
  currentOwnerPercent?: number;
  createdAt?: string;
}